+++ R/controls.R
@@ -28,7 +28,7 @@ selectizeLib <- function(bootstrap = TRUE) {
 jqueryLib <- function() {
   htmlDependency(
     name = "jquery",
-    version = "1.11.3",
+    version = "3.5.1",
     package = "crosstalk",
     src = "lib/jquery",
     script = "jquery.min.js"
@@ -258,7 +258,7 @@ inlineCheckbox <- function(id, value, label) {
 #'   according to some simple heuristics.
 #' @param animate \code{TRUE} to show simple animation controls with default
 #'   settings; \code{FALSE} not to; or a custom settings list, such as those
-#'   created using \code{\link{animationOptions}}.
+#'   created using \code{\link[shiny]{animationOptions}}.
 #' @param width The width of the slider control (see
 #'   \code{\link[htmltools]{validateCssUnit}} for valid formats)
 #' @param sep Separator between thousands places in numbers.
@@ -439,7 +439,7 @@ filter_slider <- function(id, label, sharedData, column, step = NULL,
 
   # Add animation buttons
   if (identical(animate, TRUE))
-    animate <- animationOptions()
+    animate <- shiny::animationOptions()
 
   if (!is.null(animate) && !identical(animate, FALSE)) {
     if (is.null(animate$playButton))
@@ -516,6 +516,7 @@ animation_options <- function(interval=1000,
 #' @return A \code{\link[htmltools]{browsable}} HTML element.
 #'
 #' @examples
+#' \donttest{
 #' library(htmltools)
 #'
 #' # If width is unspecified, equal widths will be used
@@ -538,6 +539,7 @@ animation_options <- function(interval=1000,
 #'   div(style = css(width="100%", height="400px", background_color="red")),
 #'   div(style = css(width="100%", height="400px", background_color="blue"))
 #' )
+#' }
 #' @export
 bscols <- function(..., widths = NA, device = c("xs", "sm", "md", "lg")) {
   device <- match.arg(device)