+++ R/ggplot2.R
@@ -81,11 +81,13 @@ selection_factor <- function(x, na.replace = c(FALSE, NA, TRUE),
 #'
 #' @export
 maintain_selection <- function(sharedData, brushId, ownerId = "") {
+  stopIfNotShiny("maintain_selection() requires the shiny package")
+
   force(sharedData)
   force(brushId)
   session <- shiny::getDefaultReactiveDomain()
 
-  observeEvent(session$input[[brushId]], {
+  shiny::observeEvent(session$input[[brushId]], {
     df <- sharedData$data(withKey = TRUE, withFilter = TRUE)
     df <- shiny::brushedPoints(df, session$input[[brushId]])
     sharedData$selection(df$key_, ownerId)