+++ NEWS.md
@@ -1,6 +1,20 @@
-## crosstalk 1.0.1.9000
+## crosstalk 1.1.1.9000
 
 * Add `selected` parameter to `filter_select`, to specify default selection.
+
+## crosstalk 1.1.1
+
+* Upgrade to jQuery v3.5.1.
+
+## crosstalk 1.1.0.1
+
+* SharedData now works even when shiny isn't installed.
+
+## crosstalk 1.1.0
+
+* `shiny` and `ggplot2` are now Suggested packages instead of Imported packages.
+
+* Upgrade to Bootstrap v3.4.1 and jQuery v3.4.1.
 
 ## crosstalk 1.0.1